+++ src/FormatFactory.cpp
@@ -48,6 +48,8 @@ namespace chemfiles {
     class MemoryBuffer;
     class Format;
 
+    extern template class Amber<AMBER_NC_RESTART>;
+    extern template class Amber<AMBER_NC_TRAJECTORY>;
     extern template class Molfile<DCD>;
     extern template class Molfile<TRJ>;
     extern template class Molfile<MOLDEN>;
@@ -61,7 +63,8 @@ static size_t find_by_extension(const std::vector<RegisteredFormat>& formats, st
 
 FormatFactory::FormatFactory() {
     // add formats in alphabetic order
-    this->add_format<AmberNetCDFFormat>();
+    this->add_format<Amber<AMBER_NC_RESTART>>();
+    this->add_format<Amber<AMBER_NC_TRAJECTORY>>();
 #ifndef CHFL_DISABLE_GEMMI
     this->add_format<CIFFormat>();
 #endif